Add rigth sidebar only to FastListTemplate
