+++ panel/template/fast/list/fast_list_template.css
@@ -1,3 +1,40 @@
 .card-margin {
   margin: 1em 0;
 }
+
+#right-sidebar-button {
+  font-size: 30px;
+  cursor: pointer;
+  opacity: 0.9;
+}
+
+#right-sidebar-button:hover {
+  opacity: 1;
+}
+
+#right-sidebar-button {
+  margin-left: 10px;
+  margin-right: 10px;
+}
+
+#right-sidebar {
+  border-right: 2px solid var(--neutral-fill-rest);
+  transition: all 0.2s cubic-bezier(0.945, 0.02, 0.27, 0.665);
+  transform-origin: center left;
+  height: calc(100vh - 76px);
+  overflow-y: auto;
+  padding: 15px 10px 5px 5px;
+}
+
+#right-sidebar.hidden {
+  border: 0px;
+  margin: 0px;
+  max-width: 0px;
+  min-width: 0px;
+  overflow: hidden;
+  padding: 0px;
+}
+
+#right-sidebar.shadow {
+  box-shadow: 2px 2px 10px silver;
+}
